+++ .idea/misc.xml
@@ -58,6 +58,26 @@
         <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_1/PhotographerBooking/app/src/main/res/layout/activity_sign_in.xml" value="0.20416666666666666" />
         <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_1/PhotographerBooking/app/src/main/res/layout/activity_sign_up.xml" value="0.20416666666666666" />
         <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_1/PhotographerBooking/app/src/main/res/layout/fragment_booking.xml" value="0.20416666666666666" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/drawable-v24/ic__star_24.xml" value="0.22447916666666667" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/drawable-v24/ic_arrow1.xml" value="0.22447916666666667" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/drawable-v24/ic_baseline_search_24.xml" value="0.22447916666666667" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/drawable-v24/round_button.xml" value="0.21666666666666667" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/drawable-v24/round_page_background.xml" value="0.22447916666666667" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/drawable/border_top_rectangle.xml" value="0.22447916666666667" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/drawable/circle.xml" value="0.22447916666666667" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/drawable/edit_text_background.xml" value="0.21666666666666667" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/drawable/ic_navigate_before_24.xml" value="0.22447916666666667" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/drawable/rectangle.xml" value="0.22447916666666667" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/drawable/rectangle_button.xml" value="0.21666666666666667" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/drawable/wallpaper_background.xml" value="0.22447916666666667" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/layout/activity_authentication.xml" value="0.20416666666666666" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/layout/activity_main.xml" value="0.20416666666666666" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/layout/activity_photographer_main.xml" value="0.20416666666666666" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/layout/activity_sign_in.xml" value="0.20416666666666666" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/layout/activity_sign_up.xml" value="0.264" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/layout/photographer_details.xml" value="0.20416666666666666" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/layout/photographer_item.xml" value="0.20416666666666666" />
+        <entry key="..\:/WORKPLACE/AndroidStudioRepo/ASM_HCI_NEW/PGbooking/app/src/main/res/layout/toolbar.xml" value="0.20416666666666666" />
         <entry key="..\:/WORKPLACE/AndroidStudioRepo/PhotographerBooking/app/src/main/res/layout/activity_main.xml" value="0.18697916666666667" />
         <entry key="..\:/WORKPLACE/AndroidStudioRepo/PhotographerBooking/app/src/main/res/layout/fragment_booking.xml" value="0.20416666666666666" />
         <entry key="..\:/WORKPLACE/AndroidStudioRepo/PhotographerBooking/app/src/main/res/layout/fragment_home.xml" value="0.1" />
