+++ app/src/main/AndroidManifest.xml
@@ -9,11 +9,11 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/Theme.PhotographerBooking">
-        <activity android:name=".MainActivity" android:exported="true" />
-        <activity android:name=".SearchActivity" android:launchMode="singleTop" />
-        <activity android:name=".home.PhotographerDetailsActivity"/>
         <activity
-            android:name=".authentication.AuthenticationActivity"
+            android:name=".authentication.SignUpActivity"
+            android:exported="false" />
+        <activity
+            android:name=".authentication.SignInActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
@@ -21,5 +21,13 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
+        <activity
+            android:name=".MainActivity"
+            android:exported="true" />
+        <activity
+            android:name=".SearchActivity"
+            android:launchMode="singleTop" />
+        <activity android:name=".home.PhotographerDetailsActivity" />
     </application>
+
 </manifest>
