+++ app/src/main/java/com/example/photographerbooking/adapter/CategoryItemsAdapter.java
@@ -11,7 +11,6 @@ import android.widget.ImageView;
 import android.widget.TextView;
 
 import androidx.annotation.NonNull;
-import androidx.appcompat.app.AppCompatActivity;
 import androidx.recyclerview.widget.RecyclerView;
 
 import com.example.photographerbooking.R;
