--- app/src/main/res/drawable-v24/ic_arrow1.xml
@@ -1,4 +1,0 @@
-<vector android:height="36dp" android:viewportHeight="512"
-    android:viewportWidth="512" android:width="36dp" xmlns:android="http://schemas.android.com/apk/res/android">
-    <path android:fillColor="#FFFFFF" android:pathData="m507.606,4.394c-3.902,-3.902 -9.632,-5.346 -14.917,-3.761l-458.881,137.665c-20.222,6.066 -33.808,24.326 -33.808,45.445 0,20.826 13.985,39.367 34.011,45.089l193.789,55.368 55.368,193.789c5.722,20.025 24.263,34.011 45.097,34.011 21.111,0 39.371,-13.586 45.438,-33.808l137.664,-458.882c1.586,-5.285 0.142,-11.014 -3.761,-14.916zM30,183.736c0,-7.761 4.994,-14.473 12.428,-16.703l401.896,-120.569 -208.756,208.756 -193.316,-55.233c-7.213,-2.062 -12.252,-8.741 -12.252,-16.251zM344.968,469.572c-2.23,7.434 -8.942,12.428 -16.711,12.428 -7.503,0 -14.182,-5.038 -16.243,-12.252l-55.233,-193.316 208.756,-208.756z"/>
-</vector>
