+++ app/src/main/res/drawable-v24/round_button.xml
@@ -3,7 +3,7 @@
     <item>
         <shape android:shape="rectangle">
             <solid android:color="@color/pink"/>
-            <corners android:radius="50dp"/>
+            <corners android:radius="15dp"/>
         </shape>
     </item>
 </selector>
