--- app/src/main/res/drawable-v24/round_page_background.xml
@@ -1,11 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item>
-        <shape android:shape="rectangle">
-            <solid android:color="@color/white"/>
-            <corners
-                android:topLeftRadius="60dp"
-                android:bottomLeftRadius="60dp"/>
-        </shape>
-    </item>
-</selector>
