+++ app/src/main/res/drawable/edit_text_background.xml
@@ -3,7 +3,8 @@
     <item>
         <shape android:shape="rectangle">
             <solid android:color="@color/white"/>
-            <corners android:radius="50dp"/>
+            <stroke android:color="@color/light_grey" android:width="3dp"/>
+            <corners android:radius="15dp"/>
         </shape>
     </item>
 </selector>
