--- app/src/main/res/drawable/tab_indicator.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
-    <item
-        android:top="-2dp"
-        android:left="-2dp"
-        android:right="-2dp"
-        android:bottom="0dp">
-
-        <shape android:shape="rectangle">
-            <solid android:color="@android:color/transparent"/>
-            <stroke android:width="1dp"/>
-        </shape>
-    </item>
-</layer-list>
