+++ app/src/main/res/values/colors.xml
@@ -8,13 +8,12 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="selected_icon">#3c415e</color>
-    <color name="count_background">#ff6f00</color>
     <color name="primary_color">#3c415e</color>
     <color name="separate_line">#80B3B3B3</color>
     <color name="one">#161853</color>
     <color name="two">#292C6D</color>
     <color name="three">#FAEDF0</color>
-
+    <color name="red">#FF0048</color>
     <color name="orange">#ff6f00</color>
     <color name="light_grey">#7FD9D9D9</color>
     <color name="pink">#EC255A</color>
