+++ app/src/main/res/values/strings.xml
@@ -4,25 +4,27 @@
     <string name="example_content">Example Content</string>
     <string name="image">image</string>
     <string name="sign_in">Sign In</string>
-    <string name="username_email">Username or email</string>
-    <string name="choose_username">Choose your username</string>
+    <string name="username">Username</string>
+    <string name="email">Email</string>
     <string name="password">Password</string>
+    <string name="enter_username">Type your username</string>
+    <string name="enter_email">Type your email</string>
+    <string name="enter_password">Type your password</string>
+    <string name="enter_confirm_password">Type confirm password</string>
     <string name="confirm_password">Confirm password</string>
     <string name="forget_password"><u>Forget password?</u></string>
     <string name="remember_me">Remember me</string>
-    <string name="no_account">No account yet? Create one</string>
-    <string name="already_account">Already have account? Login</string>
+    <string name="or_login_with">- or login with -</string>
+    <string name="no_account">No account yet?</string>
+    <string name="create_one"><u>Create one</u></string>
+    <string name="already_account">Already have an account?</string>
+    <string name="login_with_us"><u>Login with us</u></string>
     <string name="forgot_password">Forget password</string>
     <string name="type_email">Type your email</string>
     <string name="send">Send</string>
     <string name="email_instruction">No worries! Enter your email and we will send you a reset.</string>
     <string name="log_in">Login</string>
     <string name="sign_up">Sign Up</string>
-    <string name="welcome_back">Hey, Welcome Back</string>
-    <string name="let_get">Let\'s get</string>
-    <string name="started">Started!</string>
-    <string name="hello">Hello</string>
-    <string name="beautiful">Beautiful,</string>
     <string name="sign_up_instruction">Enter your information below to sign up with us</string>
     <string name="space">\u0020</string>
     <string name="app_icon">Application icon</string>
@@ -38,7 +40,7 @@
     <string name="per_session">per session</string>
     <string name="per_day">per day</string>
     <string name="per_hour">per hour</string>
-    <string name="specialization_1">Product Photography, Fashion,<br/>Business headshots, Portraits,<br/>Food<br/>Fitness Photos for Personal Trainers,<br/>Promotional Photos for Social media</string>
+    <string name="specialization_1">Product Photography, Fashion,<br />Business headshots, Portraits,<br />Food<br />Fitness Photos for Personal Trainers,<br />Promotional Photos for Social media</string>
     <string name="spot_1">Wellington Point, The LightHouse Raby Bay</string>
     <string name="chat">Chat</string>
     <string name="booking">Booking</string>
